[STEP3] Participant sealed class 추가 (Player, Dealer 상속)
