+++ src/main/kotlin/blakjack/domain/Participant.kt
@@ -1,0 +1,22 @@
+package blakjack.domain
+
+sealed class Participant(
+    val name: String
+) {
+    var cards: Cards = Cards.empty()
+        private set
+
+    val score: Int
+        get() = cards.score()
+
+    val isUnderBlackjackScore: Boolean
+        get() = cards.scoreWithAceAsOne() < BLACKJACK_SCORE
+
+    fun add(card: Card) {
+        this.cards = cards.add(card)
+    }
+
+    fun add(cards: Cards) {
+        this.cards = cards.add(cards)
+    }
+}
